python3 does not support long type anymore. Change it to int type

+++ corepy/arch/x86/lib/memory.py
@@ -71,7 +71,7 @@ class MemoryReference:
         raise Exception('Invalid scale value %s must be 1,2,4,8' % (str(scale)))
       self.scale = scale
 
-    elif isinstance(val, (int, long)):
+    elif isinstance(val, (int, int)):
       assert index is None
       assert scale == 1
       assert disp is None
